[adv-2.1] refactor: 전체 디렉토리 구조 변경(modals디렉토리를 components 디렉토리 내로 이동)
